% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argufied_functions.R
\name{argufied_functions}
\alias{argufied_functions}
\title{List argufied functions in a package or environment}
\usage{
argufied_functions(package, envir = asNamespace(package))
}
\arguments{
\item{package}{R package name. It will be loaded.}

\item{envir}{Environment to list. Defaults to the environment
of \code{package}. This means that internal and exported functions
will be listed.}
}
\value{
Data frame with columns: \code{func}, \code{assertions}.
  The latter is a logical column.
}
\description{
List argufied functions in a package or environment
}

