% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{argufy_me}
\alias{argufy_me}
\title{Function to import to run argufy on your package}
\usage{
argufy_me()
}
\description{
Call this function from the package that you want to argufy.
It does not matter where it is called from, as long as the
call is not inside other functions.
}

