% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/argufy_package.R
\name{argufy_package}
\alias{argufy_package}
\title{argufy all of the functions within a package}
\usage{
argufy_package(env = parent.frame(), ...)
}
\arguments{
\item{...}{Function arguments to modify, this allows you to argufy existing
function definitions.}

\item{package}{the package to argufy}
}
\description{
This function is best placed either in the last file sourced (\code{zzz.R}
or the last file listed in the collate directive) or in \code{.onLoad()}.
}

