% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argufy_pkgdir.R
\name{argufy_pkgdir}
\alias{argufy_pkgdir}
\title{Argufy a package that is being installed}
\usage{
argufy_pkgdir(pkg, pkg_dir, env)
}
\arguments{
\item{pkg}{Package name.}

\item{pkg_dir}{Path to source package directory.}

\item{env}{The environment containing the objects in the package.
The objects are updated in this environment.}
}
\value{
Nothing.
}
\description{
We load all help files from the package, and map the arguments
that contain assertions or coercions, to functions. (A single argument
might map to multiple functions.)
}
\details{
When we are done with all help files, we go over the functions
and inject the checks into their code.
}

