% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_enum}
\alias{as_enum}
\title{A coercion function for choice arguments in argufy}
\usage{
as_enum(arg)
}
\arguments{
\item{arg}{The argument to check and match.}
}
\value{
The matched value.
}
\description{
Check if a string scalar argument is one of the pre-specified
choices, with partial matching.
}
\seealso{
\code{\link{argufy}}, \code{\link[base]{match.arg}}
}

