% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{?}
\alias{?}
\title{Replacement for ? function}
\usage{
# ?e2
# e1?e2
}
\arguments{
\item{e1}{First argument to pass along to \code{utils::"?"}.}

\item{e2}{Second argument to pass along to \code{utils::"?"}.}
}
\description{
This function behaves identically to the existing ? function when used at
the top level, however when used within a function it is an identity function.
}

