% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ argufy }
\alias{ argufy }
\alias{ argufy }
\title{ Declerative Arguments Checks }
\description{
Declare your functions with argument checks, and 'argufy' will generate and add the checking code for you.
}
