% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rd.R
\name{rd_find}
\alias{rd_find}
\title{Find a given tag (the first one) in an Rd object}
\usage{
rd_find(rd, tag)
}
\arguments{
\item{rd}{Rd object.}

\item{tag}{Character scalar, tag to find.}
}
\value{
The tag, or NULL of not found.
}
\description{
Find a given tag (the first one) in an Rd object
}
\keyword{internal}

