% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/package.R
\name{argufy}
\alias{argufy}
\title{Add argument checks to a function}
\usage{
argufy(fun)
}
\arguments{
\item{fun}{The function to add the argument checks to.
The argument checks are specified in the original argument list
of \code{fun}. See details below.}
}
\value{
Another function that is equivalent to \code{fun}, but implements
the declared argument checks.
}
\description{
Add argument checks to a function
}

