% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{argufy}
\alias{argufy}
\title{Add argument checks to a function}
\usage{
argufy(fun, checks)
}
\description{
Function argument assertions via a concise declerative syntax.
The actual assertion code is generated automatically and inserted
at the beginning into the function.
}

