% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argufy.R
\name{argufy}
\alias{argufy}
\title{Add argument checks to a function}
\usage{
argufy(fun, assertions = NULL, coercions = NULL)
}
\arguments{
\item{fun}{Function to add assertions to.}

\item{assertions}{Character scalar, assertion expression.}

\item{coercions}{Character scalar, coercion expression.}
}
\value{
Function with assertions added.
}
\description{
Function argument assertions via a concise declerative syntax.
The actual assertion code is generated automatically and inserted
at the beginning into the function.
}
\keyword{internal}

