% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argufy.R
\name{argufy_me}
\alias{argufy_me}
\title{Function to import to run argufy on your package}
\usage{
argufy_me()
}
\description{
This function does nothing, and only exists in itself.
It can be imported into a package, and then all functions of
that package will be transformed by argufy.
}

